feat: Re-stenciling, pulling in gobox with otel upgrades
While working on portal, I ran into a dependency mismatch with otel and the vault libraries, which required upgrading the otel dependencies in gobox (https://github.com/getoutreach/gobox/pull/322, yet to land), and then upgrading the vault dependencies in vault-client (this PR).  This PR can't land until we land the gobox PR, make a release, and then update this PR with the right gobox version, but review the rest of it to make sure it's fine.