fix: return CLI output when failing to login, upgrade deps+stencil
Changes `EnsureLoggedIn` to return the error when the CLI fails
containing the output of the CLI logs, making it possible to debug why
it failed. Improved other misc error messages to include more context
for debugging purposes.

Upgraded `stencil`, cleaned up the `service.yaml` to not have null
values, and upgraded go deps for hygenic purposes.