feat: stencil and update cli interface (#99)
The purpose of this PR is to be able to use this library for accessing the production Vault server. This can be seen in the second commit of this PR. This first commit is just an auxiliary restencil.

Re-stencil the repo to update to a newer version of Go which supports log/slog.

Update the interface of EnsureLoggedIn and IsLoggedIn to remove logrus, in favor of olog/slog, and to remove the dependency on box configs, which are specific to devenv.